+++ src/lib/certpatrol-openssl.h
@@ -7,6 +7,7 @@ CertPatrolRC
 CertPatrol_OpenSSL_verify (const STACK_OF(X509) *chain,
                            const char *host, size_t host_len,
                            const char *addr, size_t addr_len,
-                           const char *proto, size_t proto_len, int port);
+                           const char *proto, size_t proto_len,
+                           uint16_t port);
 
 #endif // CERTPATROL_OPENSSL_H
